refactor: KEEP-611 share one executeWorkflowInBackground helper
The execute and webhook routes carried near-verbatim copies of the background
kick-off (start() + runId persist + error classification). Extract a single
lib/workflow/execute-in-background.ts that both call, parameterized by a log
prefix and endpoint, so the two cannot drift (the webhook copy had already lost
the credential-handling SECURITY comment).
